+++ src/app/virtualmachines/workshop/workshop.model.ts
@@ -1,6 +1,6 @@
-import { VirtualMachine } from '../virtualmachinemodels/virtualmachine';
 import { ProjectMember } from '../../projectmanagement/project_member.model';
 import { Application } from '../../applications/application.model/application.model';
+import { WorkshopVM } from './workshop-vm.model';
 
 export class Workshop {
 	application: Application;
@@ -26,21 +26,6 @@ export class Workshop {
 				for (const vm of workshop.vm_list) {
 					this.vm_list.push(new WorkshopVM(vm));
 				}
-			}
-		}
-	}
-}
-
-export class WorkshopVM {
-	vm: VirtualMachine;
-	userlogin: string;
-	elixirid: string;
-
-	constructor(workshopvm?: Partial<WorkshopVM>) {
-		Object.assign(this, workshopvm);
-		if (workshopvm) {
-			if (workshopvm.vm) {
-				this.vm = new VirtualMachine(workshopvm.vm);
 			}
 		}
 	}
